added fix for exsiting keys in emitters, start of refactor
