+++ lib/emitter/index.js
@@ -14,6 +14,7 @@ module.exports = new Base({
     require('./storage'),
     require('./off'),
     require('./set'),
+    require('./on'),
     require('./once'),
     require('./emit')
   ],
