+++ lib/emitter/on.js
@@ -1,23 +1,41 @@
 'use strict'
 const vstamp = require('vigour-stamp')
-const Base = require('vigour-base')
-const isObj = require('vigour-util/is/plainobj')
-const isNumberLike = require('vigour-util/is/numberlike')
 const listens = require('./listens')
 const ListensOnattach = listens.Attach
 const ListensOnBase = listens.Base
 const id = require('./id')
+const Base = require('vigour-base')
 
-// @todo: split this file up -- can be greatly improved
-// @todo: its dirty, its too much code, all slow, unoptmized, double
-// same here make it success
+// for off as well
+function hasKeyInOther (target, type, key) {
+  const keys = target.keys()
+  if (keys.length > 1 || keys.length && !target[type]) {
+    target.each((p, pkey) => {
+      if (pkey !== type) {
+        if (p[key]) {
+          // gogin to make this kind of stuff extremely simple
+          p.removeProperty(key)
+          // target.off(pkey, key)
+          return true // cannot have in others (wouldve been removed)
+        }
+      }
+    })
+  }
+}
 
 exports.define = {
   on (val, key, unique, stamp) {
+    var hasKey
+    if (!key) {
+      key = id(this, val)
+    } else {
+      hasKey = true
+    }
     // start at zero faster looping!
     if (!key) { key = id(this, val) }
     if (typeof val === 'function') {
       add(this, 'fn', val, key, unique, stamp)
+      if (hasKey) { hasKeyInOther(this, 'fn', key) }
     } else if (val instanceof Base) {
       if (!add(this, 'base', val, key, unique || true, stamp)) {
         if (!val.listensOnBase) {
@@ -37,6 +55,7 @@ exports.define = {
         const key = id(listensOnBase)
         listensOnBase[key] = this
         listensOnBase.addKey(key, this)
+        if (hasKey) { hasKeyInOther(this, 'base', key) }
       }
     } else if (val instanceof Array) {
       if (!add(this, 'attach', val, key, unique, stamp)) {
@@ -66,33 +85,9 @@ exports.define = {
           listensOnAttach[key] = this
           listensOnAttach.addKey(key, this)
         }
-      }
-    }
-  },
-  set (val, stamp, nocontext) {
-    if (isObj(val) && !(val instanceof Array)) {
-      set(this, val, stamp, nocontext)
-    } else {
-      this.on(val, 'val', void 0, stamp, void 0, nocontext)
-    }
-    return this
-  }
-}
-
-function set (emitter, val, stamp, nocontext) {
-  let props = emitter.properties
-  for (let key in val) {
-    if (props[key]) {
-      props[key].call(emitter, val[key], stamp, nocontext, key)
-    } else {
-      if (!emitter._id && isNumberLike(key)) {
-        // @todo: is this really nessecary? -- maybe better to check in the id generator?
-        emitter._id = Number(key)
-      }
-      if (val[key] === null) {
-        emitter.off(key)
-      } else {
-        emitter.on(val[key], key, void 0, stamp, void 0, nocontext)
+        if (hasKey) {
+          hasKeyInOther(this, 'attach', key)
+        }
       }
     }
   }
