+++ lib/emitter/set/index.js
@@ -2,8 +2,6 @@
 const isObj = require('vigour-util/is/plainobj')
 const isNumberLike = require('vigour-util/is/numberlike')
 
-exports.inject = require('./on.js')
-
 exports.define = {
   set (val, stamp, nocontext) {
     if (isObj(val) && !(val instanceof Array)) {
